third_party/blink/web_tests/virtual/plz-dedicated-worker-disabled/external/wpt/html/cross-origin-embedder-policy/multi-globals/workers-coep-report.https-expected.txt

This is a testharness.js-based test.
[FAIL] Multiple globals for Worker constructor: COEP reports
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
